crudchucvu

+++ app/Http/Controllers/api/QuanLyChamCongController.php
@@ -7,6 +7,7 @@ use App\Http\Requests\QLCC\ChamCongUpdateRequest;
 use App\Models\BangChamCong;
 use App\Models\ChiTietBangChamCong;
 use App\Models\NhanVien;
+use Illuminate\Http\Request;
 
 class QuanLyChamCongController
 {
@@ -32,7 +33,12 @@ class QuanLyChamCongController
             $nhanVien = $this->nhanVien->find($MaNV);
 
             if ($bangchamcong || $nhanVien) {
-                $chiTietBangChamCong = $this->chiTietBangChamCong->where('MaBangChamCong', $MaBCC)->where('MaNV', $MaNV)->get();
+                $chiTietBangChamCong = $this->chiTietBangChamCong->where('MaBangChamCong', $MaBCC)->where('MaNV', $MaNV)->first();
+
+                if(!$chiTietBangChamCong){
+                    return response()->json(['err' => 'Chi tiet bang cham cong khong ton tai'], 404);
+                }
+
                 return response()->json(['$chiTietBangChamCong' => $chiTietBangChamCong], 200);
             }
 
@@ -65,6 +71,7 @@ class QuanLyChamCongController
                 'TotalTime' => $request->TotalTime,
                 'Note' => $request->Note,
             ]);
+
             return response()->json(['$chiTietBangChamCong' => $chiTietBangChamCong], 200);
         } catch (\Exception $e) {
             return response()->json(['error' => $e->getMessage()], 500);
@@ -80,43 +87,52 @@ class QuanLyChamCongController
                 return response()->json(['message' => 'Chi tiet bang cham cong khong ton tai.'], 404);
             }
 
-            $this->chiTietBangChamCong->where('MaBangChamCong', $MaBCC)
-                            ->where('MaNV', $MaNV)
-                            ->update([
-                                        'MaNV' => $request->MaNV,
-                                        'MaBangChamCong' => $request->MaBangChamCong,
+            $nhanVien = $this->nhanVien->find($MaNV);
+            $BCC = $this->chamCong->find($MaBCC);
+
+            if (!$nhanVien || !$BCC) {
+                return response()->json(['error' => 'Nhan vien hoac bang cham cong khong ton tai'], 404);
+            }
+
+            $this->chiTietBangChamCong->where('MaNV', $MaNV)
+                                      ->where('MaBangChamCong', $MaBCC)
+                                      ->update([
+                                        'MaNV' => $MaNV,
+                                        'MaBangChamCong' => $MaBCC,
                                         'CheckIn' => $request->CheckIn,
                                         'CheckOut' => $request->CheckOut,
                                         'Date' => $request->Date,
                                         'TotalTime' => $request->TotalTime,
                                         'Note' => $request->Note,
                                     ]);
-            $CTBCC = $this->chiTietBangChamCong->find($MaBCC);
+
+            $CTBCC = $this->chiTietBangChamCong->where('MaNV', $MaNV)->where('MaBangChamCong',$MaBCC)->first();
             return response()->json(['chiTietBangChamCong' => $CTBCC], 200);
         } catch (\Exception $e) {
-            return response()->json([
-                'message' => 'Cap nhat that bai.',
-                'error' => $e->getMessage()
-            ], 500);
+            return response()->json(['message' => 'Cap nhat that bai.','error' => $e->getMessage()], 500);
         }
     }
 
     public function delete($MaNV, $MaBCC)
     {
-        $CTBCC = $this->chiTietBangChamCong->where('MaNV', $MaNV)->where('MaBangChamCong', $MaBCC)->first();
-
-        if (!$CTBCC) {
-            return response()->json(['message' => 'Chi tiet bang cham cong khong ton tai.'], 404);
-        }
-
         try {
+            $nhanVien = $this->nhanVien->find($MaNV);
+            $BCC = $this->chamCong->find($MaBCC);
+
+            if (!$nhanVien || !$BCC) {
+                return response()->json(['error' => 'Nhan vien hoac bang cham cong khong ton tai'], 404);
+            }
+
+            $CTBCC = $this->chiTietBangChamCong->where('MaNV', $MaNV)->where('MaBangChamCong', $MaBCC)->first();
+
+            if (!$CTBCC) {
+                return response()->json(['message' => 'Chi tiet bang cham cong khong ton tai.'], 404);
+            }
+
             $this->chiTietBangChamCong->where('MaBangChamCong', $MaBCC)->where('MaNV', $MaNV)->delete();
             return response()->json(['message' => 'Xoa thanh cong.', '$CTBCC' => $CTBCC], 200);
         } catch (\Exception $e) {
-            return response()->json([
-                'message' => 'Xoa that bai.',
-                'error' => $e->getMessage()
-            ], 500);
+            return response()->json(['message' => 'Xoa that bai.','error' => $e->getMessage()], 500);
         }
     }
 }
